Rename to once.ReturnError to satisy the linter
The linter disliked that once.OnceWithError stuttered.
So renamed it once.ReturnError
